refactor(telegram): split index.ts into preview/sync-stats/refresh modules
Mirror youtube/reddit organization: move the single-post fetch core +
fetchEventPreviewMetadata + resolveCachedExternalId into preview.ts, fetchSyncStats
into sync-stats.ts, enqueueRefreshNow into refresh.ts. index.ts stays the barrel
(adapter composition + queue topology + create hooks).

- preview.fetchEventPreviewMetadata OVERRIDES EventPreviewMetadata.externalId with
  the channelKey-based id (mirrors IG's media-id override); no channelKey resolved
  -> recognition-only (caller stores null).
- new resolveCachedExternalId re-derives the channelKey post id from our cache by
  external_url (the create-boundary trust gate, mirrors IG #70 P1) so a slug URL
  never stores a slug id.
